throwing exception in deleting existing customer logic when customer by a given id does not exists

+++ src/main/java/com/luv2code/springdemo/rest/CustomerRestController.java
@@ -64,6 +64,14 @@ public class CustomerRestController {
     @DeleteMapping("/customers/{customerId}")
     public String deleteCustomer(@PathVariable int customerId){
 
+        // check if the user exists
+        Customer customer = customerService.getCustomer(customerId);
+
+        // if not throw exception
+        if(customer == null){
+            throw new CustomerNotFoundException("Customer id not found " + customerId);
+        }
+
         customerService.deleteCustomer(customerId);
 
         return "Deleted customer id : " + customerId;
